+++ camera/ios/Plugin/ImageSaver.swift
@@ -2,9 +2,9 @@ import UIKit
 
 class ImageSaver: NSObject {
 
-    var onResult: ((Error?)->Void) = {_ in }
+    var onResult: ((Error?) -> Void) = {_ in }
 
-    init(image: UIImage, onResult:@escaping ((Error?)->Void)) {
+    init(image: UIImage, onResult:@escaping ((Error?) -> Void)) {
         self.onResult = onResult
         super.init()
         UIImageWriteToSavedPhotosAlbum(image, self, #selector(saveResult), nil)